requirements.txt
kafka-python-ng
mongodb